ref(collector): split into separate parsing file
This is a prerequisite to be able to support multiple parsers.
